pyproject.toml
[project]
name = "zalfmas-fbp"
version = "0.2.14"
description = "lib and tooling for Cap'n Proto-based FBP flows"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{ name = "Michael Berg-Mohnicke", email = "michael.berg@zalf.de" }]
dependencies = [
  "tomli>=2.2.1,<3",
  "tomli-w>=1.2.0,<2",
  "netcdf4 (>=1.7.2,<2.0.0)",
  "spotpy (>=1.6.6,<2.0.0)",
  "matplotlib (>=3.10.5,<4.0.0)",
  "pycapnp (>=2.1.0,<3.0.0)",
  "zalfmas-common (>=0.1.29,<0.2.0)",
  "zalfmas-capnp-schemas (>=0.1.36,<0.2.0)",
  "zalfmas-services (>=0.1.13,<0.2.0)",
  "pymep (>=1.0.7)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "zalfmas_fbp/run/default_configs" }]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
libsodium = ">=1.0.20,<2"

# Referencing local packages in editable mode for pixi
# [tool.pixi.pypi-dependencies]
# zalfmas-common = { path = "../mas_python_common", editable = true }
# zalfmas-capnp-schemas = { path = "../mas_capnproto_schemas", editable = true }

# Referencing local packages in editable mode for poetry
[tool.poetry.dependencies]
zalfmas-common = { path = "../mas_python_common", develop = true }
zalfmas-capnp-schemas = { path = "../mas_capnproto_schemas", develop = true }
zalfmas-services = { path = "../mas_python_services", develop = true }

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.10"

[tool.ruff]
target-version = "py310"
[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
